Split out slot searching and inventory operation utils
